add variable time units for session duration

+++ src/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, DateTime, Integer, String
+from sqlalchemy import Column, DateTime, Integer, String, Float
 
 from .database import Base
 
@@ -10,7 +10,7 @@ class TrackEntry(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     category = Column(String, nullable=False)
-    duration = Column(Integer, nullable=False)
+    duration = Column(Float, nullable=False)
     start_date = Column(DateTime, nullable=False)
     end_date = Column(DateTime, nullable=False)
     description = Column(String)
